Testing get_json
Mocking HTTP calls using unittest.mock.patch to return mock objects with json method that returns a parameterized payload passed to get_json. We verify that the get method is called exactly once per input with test_url as argumetn. We also test that the get_json is equal to test_payload